+++ flows/actions/call_resthook.go
@@ -115,10 +115,18 @@ func (a *CallResthookAction) Execute(run flows.FlowRun, step flows.Step, logModi
 
 		req.Header.Add("Content-Type", "application/json")
 
-		call, err := run.Session().Engine().Services().Webhook().Call(req, a.Resthook)
+		webhookSvc := run.Session().Engine().Services().Webhook(run.Session())
+		if webhookSvc == nil {
+			logEvent(events.NewError(errors.Errorf("no webhook provider available")))
+			return nil
+		}
+
+		call, err := webhookSvc.Call(run.Session(), req, a.Resthook)
+
 		if err != nil {
 			logEvent(events.NewError(err))
-		} else {
+		}
+		if call != nil {
 			calls = append(calls, call)
 			logEvent(events.NewWebhookCalled(call))
 		}