+++ flows/engine/services.go
@@ -1,45 +1,31 @@
 package engine
 
 import (
-	"net/http"
-
-	"github.com/nyaruka/gocommon/urns"
 	"github.com/nyaruka/goflow/flows"
-
-	"github.com/pkg/errors"
-	"github.com/shopspring/decimal"
 )
 
+// WebhookService resolves a session to a webhook provider
+type WebhookService func(flows.Session) flows.WebhookProvider
+
+// AirtimeService resolves a session to an airtime provider
+type AirtimeService func(flows.Session) flows.AirtimeProvider
+
 type services struct {
-	airtime flows.AirtimeService
-	webhook flows.WebhookService
+	webhook WebhookService
+	airtime AirtimeService
 }
 
 func newEmptyServices() *services {
 	return &services{
-		webhook: &nilWebhookService{},
-		airtime: &nilAirtimeService{},
+		webhook: func(flows.Session) flows.WebhookProvider { return nil },
+		airtime: func(flows.Session) flows.AirtimeProvider { return nil },
 	}
 }
 
-func (s *services) Webhook() flows.WebhookService {
-	return s.webhook
+func (s *services) Webhook(session flows.Session) flows.WebhookProvider {
+	return s.webhook(session)
 }
 
-func (s *services) Airtime() flows.AirtimeService {
-	return s.airtime
-}
-
-type nilWebhookService struct{}
-
-// Transfer in this case is a failure
-func (s *nilWebhookService) Call(request *http.Request, resthook string) (*flows.WebhookCall, error) {
-	return nil, errors.New("no webhook service available")
-}
-
-type nilAirtimeService struct{}
-
-// Transfer in this case is a failure
-func (s *nilAirtimeService) Transfer(session flows.Session, from urns.URN, to urns.URN, amounts map[string]decimal.Decimal) (*flows.AirtimeTransfer, error) {
-	return nil, errors.New("no airtime service available")
+func (s *services) Airtime(session flows.Session) flows.AirtimeProvider {
+	return s.airtime(session)
 }