+++ providers/webhooks/mock.go
@@ -8,33 +8,36 @@ import (
 	"github.com/nyaruka/goflow/flows"
 )
 
-type mockService struct {
-	statusCode int
-	body       string
+type mockProvider struct {
+	statusCode  int
+	contentType string
+	body        string
 }
 
-// NewMockService creates a new mock webhook service for testing
-func NewMockService(statusCode int, body string) flows.WebhookService {
-	return &mockService{
-		statusCode: statusCode,
-		body:       body,
+// NewMockProvider creates a new mock webhook provider for testing
+func NewMockProvider(statusCode int, contentType, body string) flows.WebhookProvider {
+	return &mockProvider{
+		statusCode:  statusCode,
+		contentType: contentType,
+		body:        body,
 	}
 }
 
-func (s *mockService) Call(request *http.Request, resthook string) (*flows.WebhookCall, error) {
+func (s *mockProvider) Call(session flows.Session, request *http.Request, resthook string) (*flows.WebhookCall, error) {
 	dump, err := httputil.DumpRequestOut(request, true)
 	if err != nil {
 		return nil, err
 	}
 
 	recorder := httptest.NewRecorder()
+	recorder.Header().Set("Content-Type", s.contentType)
 	recorder.WriteString(s.body)
 	recorder.Code = s.statusCode
 
 	response := recorder.Result()
 	response.Request = request
 
-	responseTrace, err := httputil.DumpResponse(response, false)
+	responseTrace, err := httputil.DumpResponse(response, true)
 	if err != nil {
 		return nil, err
 	}
@@ -51,4 +54,4 @@ func (s *mockService) Call(request *http.Request, resthook string) (*flows.Webho
 	}, nil
 }
 
-var _ flows.WebhookService = (*mockService)(nil)
+var _ flows.WebhookProvider = (*mockProvider)(nil)